Stable add CRUD for Takes, change password, and examples
Changed all DECIMAL to FLOAT in SQL due to inexplicable rounding

+++ Migrations/V2.5__InsertUpdateTakes.sql
@@ -1,4 +1,4 @@
-CREATE OR ALTER PROCEDURE insertUpdateTakes (@userid INT,@courseid INT,@grade DECIMAL)
+CREATE OR ALTER PROCEDURE insertUpdateTakes (@userid INT,@courseid INT,@grade FLOAT)
 AS
 BEGIN
 BEGIN TRANSACTION
